requirements.txt
biopython==1.72
numpy==1.15.1